src/main/java/com/leetcode/tasks/easy/NeedleInAHaystack.java
package com.leetcode.tasks.easy;

public class NeedleInAHaystack {
    public int strStr(String haystack, String needle) {
        if (needle.isEmpty()) {
            return 0;
        }
        if (haystack.length() < needle.length()) {
            return -1;
        }

        for (int i = 0; i < haystack.length(); i++) {
            if (haystack.charAt(i) == needle.charAt(0)) {
                for (int j = 0, k = i; j < needle.length() && k < haystack.length(); j++, k++) {
                    if (haystack.charAt(k) != needle.charAt(j)) {
                        break;
                    }
                    if (j == needle.length() - 1) {
                        return i;
                    }
                }
            }
        }
        return -1;
    }
}
